+++ src/buttons/CopyToClipboardButton/CopyToClipboardButton.test.jsx
@@ -1,13 +1,13 @@
 import { render, fireEvent } from '@testing-library/preact';
 import copyToClipboard from '../../utils/copy-to-clipboard';
-import CopyToClipboard from './CopyToClipboard';
+import CopyToClipboardButton from './CopyToClipboardButton';
 
 jest.mock('./../../utils/copy-to-clipboard', (value, fn) => jest.fn());
 jest.useFakeTimers();
 
-describe('UI/buttons/CopyToClipboard', () => {
+describe('UI/buttons/CopyToClipboardButton', () => {
 	it('Should call copy-to-clipboard on click', () => {
-		const { getByTestId } = render(<CopyToClipboard toCopyValue="myValue" />);
+		const { getByTestId } = render(<CopyToClipboardButton toCopyValue="myValue" />);
 		const button = getByTestId('copy-to-clipboard-button');
 		fireEvent.click(button);
 		jest.runAllTimers();
@@ -16,7 +16,7 @@ describe('UI/buttons/CopyToClipboard', () => {
 
 	it('Should call copy-to-clipboard on click with given value to be copied', () => {
 		const mockCallback = jest.fn();
-		const { getByTestId } = render(<CopyToClipboard toCopyValue="myValue" />);
+		const { getByTestId } = render(<CopyToClipboardButton toCopyValue="myValue" />);
 		const button = getByTestId('copy-to-clipboard-button');
 		fireEvent.click(button);
 		jest.runAllTimers();
@@ -26,7 +26,7 @@ describe('UI/buttons/CopyToClipboard', () => {
 
 	it('Should call copy-to-clipboard on click with callback function as second argument', () => {
 		const mockCallback = jest.fn();
-		const { getByTestId } = render(<CopyToClipboard toCopyValue="myValue" />);
+		const { getByTestId } = render(<CopyToClipboardButton toCopyValue="myValue" />);
 		const button = getByTestId('copy-to-clipboard-button');
 		fireEvent.click(button);
 		jest.runAllTimers();