+++ src/structure/Screen/Screen.jsx
@@ -3,39 +3,47 @@ import { Component } from 'preact';
 import '../../Theme/Theme';
 import CSS from './styles.css';
 
-const Screen = (props) => {
-	const [lastClick, setLastClick] = useState({x:0, y:0})
-	const viewport = { width: 0, height: 0};
+// TODO: Refactor this component from class to functional.
+// Warning: Refactoring this component to a functional one, breaks some stuffs like timeouts in CopyToClipBoardButton
+// The updates of the states are not being triggered somehow due to move this component to a functional one. (Needs investigation) 
+export default class Screen extends Component {
+	lastClick = { x: 0, y: 0 };
 
-	useEffect(() => {
-		viewport.width = window.innerWidth;
-		viewport.height = window.innerHeight;
-		window.addEventListener('resize', handleOnWindowResize);
-		return () => {
-			window.removeEventListener('resize', handleOnWindowResize);
-		};
-	}, []);
-
-	const recordClick = (event) => {
-		setLastClick({x: event.clientX, y: event.clientY});
+	viewport = {
+		width: window.innerWidth,
+		height: window.innerHeight,
 	};
 
-	const handleOnKeyboardFocus = (screenHeight, viewport) => {
-		if (props.onKeyboardFocus) { props.onKeyboardFocus(screenHeight, viewport, lastClick) };
+	recordClick({clientX, clientY}) {
+		this.lastClick.x = clientX;
+		this.lastClick.y = clientY;
+	}
+
+	onKeyboardFocus(screenHeight, viewport) {
+		if (this.props.onKeyboardFocus) this.props.onKeyboardFocus(screenHeight, viewport, this.lastClick);
+	}
+
+	onKeyboardBlur(screenHeight, viewport) {
+		if (this.props.onKeyboardFocus) this.props.onKeyboardBlur(screenHeight, viewport, this.lastClick);
+	}
+
+	onWindowResize = ({target: { innerHeight }}) => {
+		this.viewport.height > innerHeight ? this.onKeyboardFocus(innerHeight, this.viewport) : this.onKeyboardBlur(innerHeight, this.viewport);
 	};
 
-	const handleOnKeyboardBlur = (screenHeight, viewport) => {
-		if (props.onKeyboardFocus) { props.onKeyboardBlur(screenHeight, viewport, lastClick); }
-	};
+	componentDidMount() {
+		window.addEventListener('resize', this.onWindowResize);
+	}
 
-	const handleOnWindowResize = event => {
-		const {target: { innerHeight: screenSize }} = event;
-		viewport.height > (screenSize) ? handleOnKeyboardFocus(screenSize, viewport) : handleOnKeyboardBlur(screenSize, viewport);
-	};
+	componentWillUnmount() {
+		window.removeEventListener('resize', this.onWindowResize);
+	}
 
-	return  (<section className={`${CSS.screen} ${props.className || ''}`} onClick={recordClick}>
-				{props.children}
-			</section>);
+	render() {
+		const result = (<section className={`${CSS.screen} ${this.props.className || ''}`} onClick={this.recordClick}>
+			{this.props.children}
+		</section>);
+
+		return result;
+	}
 }
-
-export default Screen;